Move GitHub actions code to separate module
