+++ extensions/ql-vscode/src/remote-queries/remote-queries-manager.ts
@@ -4,7 +4,8 @@ import { CodeQLCliServer } from '../cli';
 import { ProgressCallback } from '../commandRunner';
 import { showAndLogErrorMessage, showInformationMessageWithAction } from '../helpers';
 import { Logger } from '../logging';
-import { getResultIndex, ResultIndexItem, runRemoteQuery } from './run-remote-query';
+import { runRemoteQuery } from './run-remote-query';
+import { getResultIndex, ResultIndexItem } from './gh-actions-api-client';
 import { RemoteQueriesInterfaceManager } from './remote-queries-interface';
 import { RemoteQuery } from './remote-query';
 import { RemoteQueriesMonitor } from './remote-queries-monitor';
